final working version

+++ WeatherTest/app/src/main/java/com/itis/group11801/fedotova/weathertest/net/WeatherResponse.kt
@@ -85,9 +85,9 @@ data class Weather(
 
 data class Wind(
     @SerializedName("deg")
-    var deg: Int,
+    var deg: Double,
     @SerializedName("gust")
-    var gust: Int,
+    var gust: Double,
     @SerializedName("speed")
-    var speed: Int
+    var speed: Double
 )
